requirements.txt
Flask
pyotp
qrcode
pillow
